+++ game-dyinglightthebeast/index.js
@@ -2,8 +2,8 @@
 Name: Dying Light The Beast Vortex Extension
 Structure: Basic Game
 Author: ChemBoy1
-Version: 0.1.1
-Date: 2025-09-23
+Version: 0.2.0
+Date: 2025-09-25
 ///////////////////////////////////////////*/
 
 //Import libraries
@@ -50,18 +50,28 @@ let DOWNLOAD_FOLDER = '';
 
 const PAK_ID = `${GAME_ID}-pak`;
 const PAK_NAME = "Pak Mod (Merged)";
-const PAK_PATH = path.join('ph_ft', 'source');
+//const PAK_PATH = path.join('ph_ft', 'source'); //
+const PAK_PATH = path.join('ph_ft', 'mods');
 const PAK_EXT = '.pak';
-
 const PAK_STRING = 'data';
 const PAK_IDX_START = 2;
 const PAK_IDX_END = 7;
+const VANILLA_PAKS = ['data0.pak', 'data1.pak'];
+const VANILLA_PAK_PATH = path.join('ph_ft', 'source');
 
 const ROOT_ID = `${GAME_ID}-root`;
 const ROOT_NAME = "Root Folder";
 
 const BINARIES_ID = `${GAME_ID}-binaries`;
 const BINARIES_NAME = "Binaries (Engine Injector)";
+
+const MERGER_ID = `${GAME_ID}-mergerutility`;
+const MERGER_NAME = "UTM Mod Merger Utility";
+const MERGER_PATH = 'ph_ft';
+const MERGER_EXEC = "unleashthemods.exe";
+const MERGER_EXEC_PATH = path.join(MERGER_PATH, MERGER_EXEC);
+const MERGER_PAGE_NO = 140;
+const MERGER_FILE_NO = 402;
 
 /* Config and Save paths and modtypes
 const CONFIG_ID = `${GAME_ID}-config`;
@@ -125,11 +135,17 @@ const spec = {
     }
   },
   "modTypes": [
-    /*{
+    {
       "id": PAK_ID,
       "name": PAK_NAME,
       "priority": "high",
       "targetPath": `{gamePath}\\${PAK_PATH}`
+    }, //*/
+    {
+      "id": MERGER_ID,
+      "name": MERGER_NAME,
+      "priority": "high",
+      "targetPath": `{gamePath}\\${MERGER_PATH}`
     }, //*/
     {
       "id": ROOT_ID,
@@ -166,19 +182,19 @@ const tools = [
     //defaultPrimary: true,
     parameters: PARAMETERS,
   }, //*/
-  /*{
-    id: TOOL_ID,
-    name: TOOL_NAME,
-    logo: 'tool.png',
-    executable: () => TOOL_EXEC,
+  {
+    id: MERGER_ID,
+    name: MERGER_NAME,
+    logo: 'merger.png',
+    executable: () => MERGER_EXEC_PATH,
     requiredFiles: [
-      TOOL_EXEC,
+      MERGER_EXEC_PATH,
     ],
     relative: true,
     exclusive: true,
-    //shell: true,
+    shell: true,
     //defaultPrimary: true,
-    //parameters: PARAMETERS,
+    //parameters: [],
   }, //*/
 ];
 
@@ -260,6 +276,80 @@ async function deploy(api) { //useful to deploy mods after doing some action
   return new Promise((resolve, reject) => api.events.emit('deploy-mods', (err) => err ? reject(err) : resolve()));
 }
 
+// AUTOMATIC DOWNLOADER FUNCTIONS ///////////////////////////////////////////////////
+
+//Check if Mod Merger Utility is installed
+function isMergerUtilityInstalled(api, spec) {
+  const state = api.getState();
+  const mods = state.persistent.mods[spec.game.id] || {};
+  return Object.keys(mods).some(id => mods[id]?.type === MERGER_ID);
+}
+
+//* Function to auto-download Mod Merger Utility from Nexus Mods
+async function downloadMergerUtility(api, gameSpec) {
+  let isInstalled = isMergerUtilityInstalled(api, gameSpec);
+  if (!isInstalled) {
+    const MOD_NAME = MERGER_NAME;
+    const MOD_TYPE = MERGER_ID;
+    const NOTIF_ID = `${MOD_TYPE}-installing`;
+    let FILE_ID = MERGER_FILE_NO;  //If using a specific file id because "input" below gives an error
+    const PAGE_ID = MERGER_PAGE_NO;
+    const GAME_DOMAIN = GAME_ID;
+    api.sendNotification({ //notification indicating install process
+      id: NOTIF_ID,
+      message: `Installing ${MOD_NAME}`,
+      type: 'activity',
+      noDismiss: true,
+      allowSuppress: false,
+    });
+    if (api.ext?.ensureLoggedIn !== undefined) { //make sure user is logged into Nexus Mods account in Vortex
+      await api.ext.ensureLoggedIn();
+    }
+    try {
+      let FILE = FILE_ID; //use the FILE_ID directly for the correct game store version
+      let URL = `nxm://${GAME_DOMAIN}/mods/${PAGE_ID}/files/${FILE}`;
+      try { //get the mod files information from Nexus
+        const modFiles = await api.ext.nexusGetModFiles(GAME_DOMAIN, PAGE_ID);
+        const fileTime = () => Number.parseInt(input.uploaded_time, 10);
+        const file = modFiles
+          .filter(file => file.category_id === 1)
+          .sort((lhs, rhs) => fileTime(lhs) - fileTime(rhs))[0];
+        if (file === undefined) {
+          throw new util.ProcessCanceled(`No ${MOD_NAME} main file found`);
+        }
+        FILE = file.file_id;
+        URL = `nxm://${GAME_DOMAIN}/mods/${PAGE_ID}/files/${FILE}`;
+      } catch (err) { // use defined file ID if input is undefined above
+        FILE = FILE_ID;
+        URL = `nxm://${GAME_DOMAIN}/mods/${PAGE_ID}/files/${FILE}`;
+      } //
+      const dlInfo = { //Download the mod
+        game: GAME_DOMAIN,
+        name: MOD_NAME,
+      };
+      const dlId = await util.toPromise(cb =>
+        api.events.emit('start-download', [URL], dlInfo, undefined, cb, undefined, { allowInstall: false }));
+      const modId = await util.toPromise(cb =>
+        api.events.emit('start-install-download', dlId, { allowAutoEnable: false }, cb));
+      const profileId = selectors.lastActiveProfileForGame(api.getState(), gameSpec.game.id);
+      const batched = [
+        actions.setModsEnabled(api, profileId, [modId], true, {
+          allowAutoDeploy: true,
+          installed: true,
+        }),
+        actions.setModType(gameSpec.game.id, modId, MOD_TYPE), // Set the mod type
+      ];
+      util.batchDispatch(api.store, batched); // Will dispatch both actions
+    } catch (err) { //Show the user the download page if the download, install process fails
+      const errPage = `https://www.nexusmods.com/${GAME_DOMAIN}/mods/${PAGE_ID}/files/?tab=files`;
+      api.showErrorNotification(`Failed to download/install ${MOD_NAME}`, err);
+      util.opn(errPage).catch(() => null);
+    } finally {
+      api.dismissNotification(NOTIF_ID);
+    }
+  }
+} //*/
+
 // MOD INSTALLER FUNCTIONS ///////////////////////////////////////////////////
 
 //Test for .pak files (in mod merger)
@@ -280,7 +370,8 @@ function testPak(files, gameId) {
   });
 }
 
-function installPak(api, files) {
+/*Install pak files (merger function version)
+function installPak(api, files, fileName) {
   const rootCandidate = files.find(file => file.toLowerCase().split(path.sep).includes('ph_ft'));
   const idx = rootCandidate !== undefined
     ? rootCandidate.toLowerCase().split(path.sep).findIndex(seg => seg === 'ph_ft')
@@ -375,6 +466,166 @@ function installPak(api, files) {
   const prom = hasVariants ? queryVariant : Promise.resolve;
   return prom()
     .then(() => Promise.resolve({ instructions: generateInstructions() }));
+} //*/
+
+//*Install pak files (Mod Merger Utility version)
+function installPak(api, files, fileName) {
+  const rootCandidate = files.find(file => file.toLowerCase().split(path.sep).includes('ph_ft'));
+  const idx = rootCandidate !== undefined
+    ? rootCandidate.toLowerCase().split(path.sep).findIndex(seg => seg === 'ph_ft')
+    : 0;
+  
+  const MOD_NAME = path.basename(fileName);
+  const MOD_FOLDER = MOD_NAME.replace(/[\.]*(installing)*(zip)*/gi, '');
+
+  let hasVariants = false;
+  const pakFiles = files.reduce((accum, iter) => {
+    if (path.extname(iter) === '.pak') {
+      const exists = accum[path.basename(iter)] !== undefined;
+      if (exists) {
+        hasVariants = true;
+      }
+      accum[path.basename(iter)] = exists
+        ? accum[path.basename(iter)].concat(iter)
+        : [iter];
+    }
+    return accum;
+  }, {});
+
+  let filtered = files;
+  const queryVariant = () => {
+    const paks = Object.keys(pakFiles).filter(key => pakFiles[key].length > 1);
+    return Promise.map(paks, pakFile => {
+      return api.showDialog('question', 'Choose Variant', {
+        text: 'This mod has several variants for "{{pak}}" - please '
+            + 'choose the variant you wish to install. (You can choose a '
+            + 'different variant by re-installing the mod)',
+        choices: pakFiles[pakFile].map((iter, idx) => ({ 
+          id: iter,
+          text: iter,
+          value: idx === 0,
+        })),
+        parameters: {
+          pak: pakFile,
+        },
+      }, [
+        { label: 'Cancel' },
+        { label: 'Confirm' },
+      ]).then(res => {
+        if (res.action === 'Confirm') {
+          const choice = Object.keys(res.input).find(choice => res.input[choice]);
+          filtered = filtered.filter(file => (path.extname(file) !== PAK_EXT)
+            || ((path.basename(file) === pakFile) && file.includes(choice))
+            || (path.basename(file) !== pakFile));
+          return Promise.resolve();
+        } else {
+          return new util.UserCanceled();
+        }
+      });
+    })
+  };
+  const generateInstructions = () => {
+    const fileInstructions = filtered.reduce((accum, iter) => {
+      if (!iter.endsWith(path.sep)) {
+        accum.push({
+          type: 'copy',
+          source: iter,
+          destination: path.join(MOD_FOLDER, path.basename(iter)),
+        });
+      }
+      return accum;
+    }, []);
+    const instructions = [{ 
+      type: 'setmodtype',
+      value: PAK_ID,
+    }].concat(fileInstructions);
+    return instructions;
+  }
+
+  const prom = hasVariants ? queryVariant : Promise.resolve;
+  return prom()
+    .then(() => Promise.resolve({ instructions: generateInstructions() }));
+} //*/
+
+//* Install paks in zips for Mod Merger Utility
+async function installZipContent(files, destinationPath) {
+  const zipFiles = files.filter(file => ['.zip', '.7z', '.rar'].includes(path.extname(file)));
+  const setModTypeInstruction = { type: 'setmodtype', value: PAK_ID };
+  if (zipFiles.length > 0) { // If it's a double zip, we don't need to repack. 
+    const instructions = zipFiles.map(file => {
+      return {
+        type: 'copy',
+        source: file,
+        destination: path.basename(file),
+      }
+    });
+    instructions.push(setModTypeInstruction);
+    return Promise.resolve({ instructions });
+  }
+  else { // Repack the ZIP
+    const szip = new util.SevenZip();
+    const MOD_NAME = path.basename(destinationPath);
+    const ZIP_NAME = MOD_NAME.replace(/[\-]*[\.]*[\d]*[ ]*(installing)*(zip)*/gi, '');
+    const archiveName = ZIP_NAME + '.zip';
+    const archivePath = path.join(destinationPath, archiveName);
+    const rootRelPaths = await fs.readdirAsync(destinationPath);
+    await szip.add(archivePath, rootRelPaths.map(relPath => path.join(destinationPath, relPath)), { raw: ['-r'] });
+    const instructions = [{
+      type: 'copy',
+      source: archiveName,
+      destination: path.basename(archivePath),
+    }];
+    instructions.push(setModTypeInstruction);
+    return Promise.resolve({ instructions });
+  }
+}
+
+const Bluebird = require('bluebird');
+//convert installer functions to Bluebird promises
+function toBlue(func) {
+  return (...args) => Bluebird.Promise.resolve(func(...args));
+} //*/
+
+//Installer test for Hotfix Merger files
+function testMergerUtility(files, gameId) {
+  const isMod = files.some(file => (path.basename(file).toLowerCase() === MERGER_EXEC));
+  let supported = (gameId === spec.game.id) && isMod;
+
+  /* Test for a mod installer.
+  if (supported && files.find(file =>
+    (path.basename(file).toLowerCase() === 'moduleconfig.xml') &&
+    (path.basename(path.dirname(file)).toLowerCase() === 'fomod'))) {
+    supported = false;
+  } //*/
+
+  return Promise.resolve({
+    supported,
+    requiredFiles: [],
+  });
+}
+
+//Installer install Hotfix Merger files
+function installMergerUtility(files) {
+  const MOD_TYPE = MERGER_ID;
+  const modFile = files.find(file => (path.basename(file).toLowerCase() === MERGER_EXEC));
+  const idx = modFile.indexOf(path.basename(modFile));
+  const rootPath = path.dirname(modFile);
+  const setModTypeInstruction = { type: 'setmodtype', value: MOD_TYPE };
+
+  // Remove directories and anything that isn't in the rootPath.
+  const filtered = files.filter(file =>
+    ((file.indexOf(rootPath) !== -1) && (!file.endsWith(path.sep)))
+  );
+
+  const instructions = filtered.map(file => {
+    return {
+      type: 'copy',
+      source: file,
+      destination: path.join(file.substr(idx)),
+    };
+  });
+  instructions.push(setModTypeInstruction);
+  return Promise.resolve({ instructions });
 }
 
 //Installer test for Root folder files
@@ -519,24 +770,6 @@ function mergeOperation(api, filePath, mergeDir) {
     .then(() => fs.moveAsync(zipFile, mergeFilePath, { overwrite: true }));
 }
 
-const requestDeployment = (context) => {
-  context.api.store.dispatch(actions.setDeploymentNecessary(GAME_ID, true));
-  context.api.sendNotification({
-    id: `${GAME_ID}-deployrequest`,
-    type: 'warning',
-    message: 'Deployment is needed',
-    allowSuppress: true,
-    actions: [
-      {
-        title: 'Deploy',
-        action: () => context.api.events.emit('deploy-mods', (err) => {
-          console.warn(`Error deploying mods \n${err}`)
-        })
-      }
-    ],
-  });
-};
-
 // MAIN FUNCTIONS ///////////////////////////////////////////////////////////////
 
 //Setup function
@@ -547,6 +780,14 @@ async function setup(discovery, api, gameSpec) {
   STAGING_FOLDER = selectors.installPathForGame(state, GAME_ID);
   DOWNLOAD_FOLDER = selectors.downloadPathForGame(state, GAME_ID);
   // ASYNC CODE //////////////////////////////////////////
+  await downloadMergerUtility(api, gameSpec);
+  const MERGER_FOLDER_OLD = path.join(STAGING_FOLDER, '__merged.dyinglightthebeast-pak');
+  try {
+    await fs.statAsync(MERGER_FOLDER_OLD);
+    await fs.removeAsync(MERGER_FOLDER_OLD);
+  } catch (err) {
+    //do nothing
+  } //*/
   return fs.ensureDirWritableAsync(path.join(GAME_PATH, MOD_PATH_DEFAULT));
 }
 
@@ -573,21 +814,23 @@ function applyGame(context, gameSpec) {
     }, (game) => pathPattern(context.api, game, type.targetPath), () => Promise.resolve(false), { name: type.name });
   });
 
-  //register mod types explicitly
-    context.registerModType(PAK_ID, 25, //id, priority
-      (gameId) => {
-        var _a;
-        return (gameId === GAME_ID) && !!((_a = context.api.getState().settings.gameMode.discovered[gameId]) === null || _a === void 0 ? void 0 : _a.path);
-      }, //isSupported - Is this mod for this game
-      (game) => pathPattern(context.api, game, `{gamePath}\\${PAK_PATH}`), //getPath - mod install location
-      () => Promise.resolve(false), //test - is installed mod of this type
-      {
-        name: PAK_NAME,
-      } //options
-    );
+  /*register mod types explicitly
+  context.registerModType(PAK_ID, 25, //id, priority
+    (gameId) => {
+      var _a;
+      return (gameId === GAME_ID) && !!((_a = context.api.getState().settings.gameMode.discovered[gameId]) === null || _a === void 0 ? void 0 : _a.path);
+    }, //isSupported - Is this mod for this game
+    (game) => pathPattern(context.api, game, `{gamePath}\\${PAK_PATH}`), //getPath - mod install location
+    () => Promise.resolve(false), //test - is installed mod of this type
+    {
+      name: PAK_NAME,
+    } //options
+  ); //*/
 
   //register mod installers
-  context.registerInstaller(PAK_ID, 25, testPak, (files) => installPak(context.api, files));
+  //context.registerInstaller(PAK_ID, 25, testPak, (files, fileName) => installPak(context.api, files, fileName));
+  context.registerInstaller(PAK_ID, 25, toBlue(testPak), toBlue(installZipContent));
+  context.registerInstaller(MERGER_ID, 27, testMergerUtility, installMergerUtility);
   //context.registerInstaller(CONFIG_ID, 43, testConfig, installConfig);
   //context.registerInstaller(SAVE_ID, 45, testSave, installSave);
   context.registerInstaller(ROOT_ID, 47, testRoot, installRoot);
@@ -615,7 +858,7 @@ function applyGame(context, gameSpec) {
 //main function
 function main(context) {
   applyGame(context, spec);
-  //*merger for pak mods
+  /* merger for pak mods
   context.registerMerge(
     (game, discovery) => mergeTest(context.api, game, discovery),
     (filePath, mergeDir) => mergeOperation(context.api, filePath, mergeDir),
@@ -623,13 +866,132 @@ function main(context) {
   ); //*/
 
   context.once(() => { // put code here that should be run (once) when Vortex starts up
-    /*context.api.onAsync('did-deploy', async (profileId, deployment) => {
+    context.api.onAsync('did-deploy', async (profileId, deployment) => {
       const lastActiveProfile = selectors.lastActiveProfileForGame(context.api.getState(), GAME_ID);
       if (profileId !== lastActiveProfile) return;
-      context.api.dismissNotification('redundant-mods'); // Because we create a merged mod when deploying, Vortex thinks that all mods have duplicates and are redundant
+      await didDeploy(context.api);
+      return deployNotify(context.api);
+    }); //*/
+    context.api.onAsync('did-purge', async (profileId, deployment) => {
+      const lastActiveProfile = selectors.lastActiveProfileForGame(context.api.getState(), GAME_ID);
+      if (profileId !== lastActiveProfile) return;
+      return didPurge(context.api);
     }); //*/
   });
   return true;
+}
+
+async function didDeploy(api) {
+  const state = api.getState();
+  GAME_PATH = getDiscoveryPath(api);
+  const PAK_DIRECTORY = path.join(GAME_PATH, VANILLA_PAK_PATH);
+  let FILES = await fs.readdirAsync(PAK_DIRECTORY);
+  
+  try { //clear pak non-vanilla pak files
+    FILES = FILES.filter(file => 
+      path.extname(file).toLowerCase() === PAK_EXT &&
+      !VANILLA_PAKS.includes(path.basename(file))
+    );
+    //log('warn', `Removing pak files on deploy: ${FILES.join(', ')}`);
+    FILES.forEach(async file => {
+      await fs.removeAsync(path.join(PAK_DIRECTORY, file));
+    });
+  } catch (err) {
+    log('error', `Failed to remove merged pak files: ${FILES.join(', ')} - ${err.message}`);
+  }
+
+  //runModManager(api); //run merger executable once files are cleared - not executed for technical reasons
+  return Promise.resolve();
+}
+
+async function didPurge(api) {
+  const state = api.getState();
+  GAME_PATH = getDiscoveryPath(api);
+  const PAK_DIRECTORY = path.join(GAME_PATH, VANILLA_PAK_PATH);
+  let FILES = await fs.readdirAsync(PAK_DIRECTORY);
+  
+  try { //clear pak non-vanilla pak files
+    FILES = FILES.filter(file => 
+      path.extname(file).toLowerCase() === PAK_EXT &&
+      !VANILLA_PAKS.includes(path.basename(file))
+    );
+    //log('warn', `Removing pak files on purge: ${FILES.join(', ')}`);
+    FILES.forEach(async file => {
+      await fs.removeAsync(path.join(PAK_DIRECTORY, file));
+    });
+  } catch (err) {
+    log('error', `Failed to remove merged pak files: ${FILES.join(', ')} - ${err.message}`);
+  }
+  return Promise.resolve();
+}
+
+//Notify User to run Mod Merger Utility after deployment
+function deployNotify(api) {
+  const NOTIF_ID = `${GAME_ID}-deploy`;
+  const MOD_NAME = MERGER_NAME;
+  const MESSAGE = `Use ${MOD_NAME} to Install Mods`;
+  api.sendNotification({
+    id: NOTIF_ID,
+    type: 'warning',
+    message: MESSAGE,
+    allowSuppress: true,
+    actions: [
+      {
+        title: 'Run Merger',
+        action: (dismiss) => {
+          runModManager(api);
+          dismiss();
+        },
+      },
+      {
+        title: 'More',
+        action: (dismiss) => {
+          api.showDialog('question', MESSAGE, {
+            text: `For most mods, you must use ${MOD_NAME} to install the mod to the game files after installing with Vortex.\n`
+                + `Mods to install with ${MOD_NAME} will be found at this folder: "${HOTFIX_PATH}".\n`
+                + `Use the included tool to launch ${MOD_NAME} (button below or in "Dashboard" tab).\n`
+                + `You can open the Hotfix Merger WebUI using the button below, or using the button within the folder icon on the Mods toolbar.\n`
+          }, [
+            {
+              label: 'Run Merger', action: () => {
+                runModManager(api);
+                dismiss();
+              }
+            },
+            { label: 'Continue', action: () => dismiss() },
+            {
+              label: 'Never Show Again', action: () => {
+                api.suppressNotification(NOTIF_ID);
+                dismiss();
+              }
+            },
+          ]);
+        },
+      },
+    ],
+  });
+}
+
+function runModManager(api) {
+  const TOOL_ID = MERGER_ID;
+  const TOOL_NAME = MERGER_NAME;
+  const state = api.store.getState();
+  const tool = util.getSafe(state, ['settings', 'gameMode', 'discovered', GAME_ID, 'tools', TOOL_ID], undefined);
+
+  try {
+    const TOOL_PATH = tool.path;
+    if (TOOL_PATH !== undefined) {
+      return api.runExecutable(TOOL_PATH, [], { suggestDeploy: false, shell: true })
+        .catch(err => api.showErrorNotification(`Failed to run ${TOOL_NAME}`, err,
+          { allowReport: ['EPERM', 'EACCESS', 'ENOENT'].indexOf(err.code) !== -1 })
+        );
+    }
+    else {
+      return api.showErrorNotification(`Failed to run ${TOOL_NAME}`, `Path to ${TOOL_NAME} executable could not be found. Ensure ${TOOL_NAME} is installed through Vortex.`);
+    }
+  } catch (err) {
+    return api.showErrorNotification(`Failed to run ${TOOL_NAME}`, err, { allowReport: ['EPERM', 'EACCESS', 'ENOENT'].indexOf(err.code) !== -1 });
+  }
 }
 
 //export to Vortex